+++ pp/difftest.py
@@ -1,3 +1,4 @@
+"""GDS regression test."""
 import pathlib
 
 from lytest.kdb_xor import GeometryDifference, run_xor
@@ -10,7 +11,8 @@ cwd = pathlib.Path.cwd()
 
 
 def difftest(component: Component) -> None:
-    """Runs an XOR over a component and makes boolean comparison with a GDS reference.
+    """Avoids GDS regressions tests on the GeometryDifference.
+    Runs an XOR over a component and makes boolean comparison with a GDS reference.
     If it runs for the fist time it just stores the GDS reference.
     raises GeometryDifference if there are differences and show differences in klayout.
     """